data_pipeline/requirements.txt
# Biz-Flow 데이터 파이프라인 의존성
requests==2.31.0
google-generativeai==0.5.4
